+++ demo/admin/webpack.config.ts
@@ -21,15 +21,6 @@ const config = ({ production }: IEnvironment): webpack.Configuration => {
             },
             hash: true,
         }),
-        new HtmlWebpackPlugin({
-            template: require.resolve("@comet/cms-admin/access-token-service-worker.ejs"),
-            templateParameters: {
-                API_URL: production ? "$API_URL" : process.env.API_URL,
-            },
-            filename: "access-token-service-worker.js",
-            inject: false,
-            minify: false,
-        }),
     ];
     if (production) {
         plugins.push(
@@ -143,11 +134,6 @@ const config = ({ production }: IEnvironment): webpack.Configuration => {
             port: Number(process.env.ADMIN_PORT || 8001),
             allowedHosts: "all",
             compress: true,
-            client: {
-                webSocketURL: {
-                    port: Number(process.env.PROXY_PORT),
-                },
-            },
             historyApiFallback: true,
             headers: {
                 "Access-Control-Allow-Origin": "*",