+++ ecosystem-demo-api.config.js
@@ -19,17 +19,5 @@ module.exports = {
             namespace: "comet-cms",
             autorestart: true,
         },
-        {
-            name: "comet-demo-proxy",
-            script: "node proxy.js",
-            namespace: "comet-cms",
-            autorestart: true,
-        },
-        {
-            name: "comet-demo-idp",
-            script: "node mock-idp.js",
-            namespace: "comet-cms",
-            autorestart: true,
-        },
     ],
 };